Begins conversion of AABBs to convex collision meshes
Collision resolution is still jerky.  But only after I changed collision meshes
to store the position as a field rather than translating every single vertex by
a position and not storing the position as a field.

The plan now is to fix the jittery collisions, and then to speed things up,
store an AABB with each mesh to do a quick initial detection pass.

Additional:
    - Meshes in OBJs with a "CM_" prefix define a collision mesh for the mesh
      with the name that has the corresponding suffix.  Name matching as just
      described has not actually been implemented yet, because there is
      currently only support for defining a single mesh and, optionally, a
      collision mesh.